% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_map.R
\name{kde_map}
\alias{kde_map}
\title{Kernel Density Estimation and Heat Map Generation for Crime Incidents}
\usage{
kde_map(data)
}
\arguments{
\item{data}{Data frame of crime or RMS data. See provided Chicago Data Portal
example for reference}
}
\value{
A \pkg{Leaflet} map with three layers: an ESRI base-map, all crime
    incidents plotted (with incident info pop-up windows), and a kernel
    density estimate of those points.
}
\description{
This function computes a kernel density estimate of crime
    incident locations and returns a Leaflet map of the incidents. The data
    is based on the Chicago Police Department RMS structure and populates
    pop-up windows with the incident location for each incident.
}
\examples{
\donttest{
#Using provided dataset from Chicago Data Portal:
crimes <- test:::crimes
library("leaflet") # needed to install basemap providers
kde_map(crimes)}
}
\author{
Jamie Spaulding, Keith Morris
}
\keyword{dynamic}
\keyword{hplot}
\keyword{methods}
\keyword{spatial}
