% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_daily_decomp.R
\name{ts_daily_decomp}
\alias{ts_daily_decomp}
\title{Time Series Forecast and Decomposition for Daily Crime Data}
\usage{
ts_daily_decomp(data)
}
\arguments{
\item{data}{Data frame of crime or RMS data. See provided Chicago Data Portal
example for reference}
}
\value{
Returns a list of four plots: the overall crime trend with forecast;
    the daily seasonality; the weekly seasonality; and the yearly seasonality
    components.
}
\description{
Plot the components of forecast generated using the \pkg{prophet}
    which includes the overall crime trend and the daily, weekly, and yearly
    seasonality components.
}
\examples{
\donttest{
#Using provided dataset from Chicago Data Portal:
crimes <- test:::crimes
library(prophet)
ts_daily_decomp(crimes)}
}
\author{
Jamie Spaulding, Keith Morris
}
\keyword{ts}
